+++ packages/calcite-components/src/components/sort-handle/sort-handle.tsx
@@ -63,7 +63,6 @@ export class SortHandle extends LitElement implements LoadableComponent, Interac
    * Made into a prop for testing purposes only.
    *
    * @private
-   * @readonly
    */
   @property() messages = useT9n<typeof T9nStrings>({ blocking: true });