+++ packages/calcite-components/src/custom-theme.stories.ts
@@ -38,6 +38,7 @@ import { pagination } from "./custom-theme/pagination";
 import { popover, popoverTokens } from "./custom-theme/popover";
 import { progress, progressTokens } from "./custom-theme/progress";
 import { segmentedControl } from "./custom-theme/segmented-control";
+import { rating, ratingTokens } from "./custom-theme/rating";
 import { slider, sliderTokens } from "./custom-theme/slider";
 import { switchTokens } from "./custom-theme/switch";
 import { tabs, tabsBordered, tabsTokens } from "./custom-theme/tabs";
@@ -48,6 +49,7 @@ import { navigationLogoTokens, navigationLogos } from "./custom-theme/navigation
 import { navigationUserTokens, navigationUsers } from "./custom-theme/navigation-user";
 import { tileTokens, tile } from "./custom-theme/tile";
 import { navigationTokens, navigation } from "./custom-theme/navigation";
+import { menuItem, menuItemTokens } from "./custom-theme/menu-item";
 
 const globalTokens = {
   calciteColorBrand: "#007ac2",
@@ -133,8 +135,11 @@ const kitchenSink = (args: Record<string, string>, useTestValues = false) =>
         ${avatarInitials} ${avatarThumbnail} ${progress} ${handle} ${textArea} ${popover} ${tile} ${tooltip}
         ${comboboxItem}
       </div>
-      <div class="demo-column">${navigation} ${navigationLogos} ${navigationUsers} ${blockSection} ${block}</div>
+      <div class="demo-column">
+        ${navigation} ${navigationLogos} ${navigationUsers} ${blockSection} ${block} ${rating}
+      </div>
       <div class="demo-column"><div class="demo-column">${alert}</div></div>
+      <div class="demo-column">${menuItem}</div>
     </div>
   </div>`;
 
@@ -167,12 +172,14 @@ const componentTokens = {
   ...noticeTokens,
   ...popoverTokens,
   ...progressTokens,
+  ...ratingTokens,
   ...sliderTokens,
   ...switchTokens,
   ...tabsTokens,
   ...textAreaTokens,
   ...tileTokens,
   ...tooltipTokens,
+  ...menuItemTokens,
 };
 
 export default {